HttpClient Structure refactored
