+++ Application/Application.csproj
@@ -16,8 +16,4 @@
     <ProjectReference Include="..\Infrastructure\Infrastructure.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Dtos\Mappers\" />
-  </ItemGroup>
-
 </Project>
