--- Application/Dtos/OrderDtos/CompleteOrderRequestDto.cs
@@ -1,5 +1,0 @@
-﻿namespace Application.Dtos.OrderDtos
-{
-    public record CompleteOrderRequestDto(string orderId);
-    
-}
