--- Application/Dtos/OrderDtos/OrderResponseDto.cs
@@ -1,6 +1,0 @@
-﻿namespace Application.Dtos.OrderDtos
-{
-    public record OrderResponseDto(bool success,
-    string message,
-    PreorderDataDto data);
-}
