--- Application/Dtos/OrderDtos/PreOrderDto.cs
@@ -1,9 +1,0 @@
-﻿namespace Application.Dtos.OrderDtos
-{
-    public record PreOrderDto(
-    string orderId,
-    decimal amount,
-    DateTime timestamp,
-    string status
-);
-}
