--- Application/Dtos/OrderDtos/PreOrderRequestDto.cs
@@ -1,5 +1,0 @@
-﻿namespace Application.Dtos.OrderDtos
-{
-    public record PreOrderRequestDto(string orderId, decimal amount);
-    
-}
