--- Application/Dtos/OrderDtos/PreorderDataDto.cs
@@ -1,7 +1,0 @@
-﻿namespace Application.Dtos.OrderDtos
-{
-    public record PreorderDataDto(
-    PreOrderDto preOrder,
-    UpdatedBalanceDto updatedBalance
-);
-}
