--- Application/Dtos/ProductDtos/GetProductsClientResponseDto.cs
@@ -1,4 +1,0 @@
-﻿namespace Application.Dtos.ProductDtos
-{
-    public record GetProductsClientResponseDto(bool success, List<ProductDto> data);
-}
