--- Application/Dtos/ProductDtos/ProductDto.cs
@@ -1,5 +1,0 @@
-﻿
-namespace Application.Dtos.ProductDtos
-{
-    public record ProductDto(string id, string name, string description, decimal price,string currency,string category,int stock);
-}
