+++ Dtos/InfrastructureDtos/HttpClientDtos/BalanceManagementClientDtos/OrderDtos/CompleteOrderRequestDto.cs
@@ -1,0 +1,4 @@
+﻿namespace Dtos.InfrastructureDtos.HttpClientDtos.BalanceManagementClientDtos.OrderDtos
+{
+    public record CompleteOrderRequestDto(string orderId);
+}
