+++ Dtos/InfrastructureDtos/HttpClientDtos/BalanceManagementClientDtos/OrderDtos/OrderResponseDto.cs
@@ -1,0 +1,6 @@
+﻿namespace Dtos.InfrastructureDtos.HttpClientDtos.BalanceManagementClientDtos.OrderDtos
+{
+    public record OrderResponseDto(bool success,
+   string message,
+   PreorderDataDto data);
+}
