+++ Dtos/InfrastructureDtos/HttpClientDtos/BalanceManagementClientDtos/OrderDtos/PreOrderDto.cs
@@ -1,0 +1,9 @@
+﻿namespace Dtos.InfrastructureDtos.HttpClientDtos.BalanceManagementClientDtos.OrderDtos
+{
+    public record PreOrderDto(
+        string orderId,
+        decimal amount,
+        DateTime timestamp,
+        string status
+        );
+}
