+++ Dtos/InfrastructureDtos/HttpClientDtos/BalanceManagementClientDtos/OrderDtos/PreOrderRequestDto.cs
@@ -1,0 +1,4 @@
+﻿namespace Dtos.InfrastructureDtos.HttpClientDtos.BalanceManagementClientDtos.OrderDtos
+{
+    public record PreOrderRequestDto(string orderId, decimal amount);
+}
