+++ Dtos/InfrastructureDtos/HttpClientDtos/BalanceManagementClientDtos/OrderDtos/PreorderDataDto.cs
@@ -1,0 +1,7 @@
+﻿namespace Dtos.InfrastructureDtos.HttpClientDtos.BalanceManagementClientDtos.OrderDtos
+{
+    public record PreorderDataDto(
+  PreOrderDto preOrder,
+  UpdatedBalanceDto updatedBalance
+);
+}
