+++ Dtos/InfrastructureDtos/HttpClientDtos/BalanceManagementClientDtos/OrderDtos/UpdatedBalanceDto.cs
@@ -1,0 +1,10 @@
+﻿namespace Dtos.InfrastructureDtos.HttpClientDtos.BalanceManagementClientDtos.OrderDtos
+{
+    public record UpdatedBalanceDto(
+        Guid userId,
+        decimal totalBalance,
+        decimal availableBalance,
+        decimal blockedBalance,
+        string currency,
+        DateTime lastUpdated);
+}
