+++ Dtos/InfrastructureDtos/HttpClientDtos/BalanceManagementClientDtos/ProductDtos/GetProductsClientResponseDto.cs
@@ -1,0 +1,4 @@
+﻿namespace Dtos.InfrastructureDtos.HttpClientDtos.BalanceManagementClientDtos.ProductDtos
+{
+    public record GetProductsClientResponseDto(bool success, List<ProductDto> data);
+}
